+++ examples/geotiff_download/lib/api2/ta_task.py
@@ -20,7 +20,7 @@ class TerrAvionAPI2Task:
             result = r.json()
             return result
         else:
-            self.log.debug('error:' + str(r.status_code))
+            self.log.debug('error: %s', str(r.status_code))
             self.log.debug(r.text)
             self.log.debug('-------------------------------------------------------')